+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Fixed
 * **core:** Dropdown menu sluit niet meer bij selecteren van item ([#1580](https://github.com/dso-toolkit/dso-toolkit/issues/1580))
+* **css + dso-toolkit + sources:** Shopping Cart 'non-collapsable' variant toevoegen ([#1581](https://github.com/dso-toolkit/issues/1581))
 
 ## 38.0.0